replaced uuid with hash
since there are two files per exam (texts.csv and actual.csv), the scrubbing of the student IDs had to be done in both place. The original commit  did this in one place only. So the method was changed to a hash w/ random seed so IDs could be changed consistently across files.