fixed bug, updated output
The `grade()` function was being passed all texts, not just those for which there were human. Although this is likely to provide better results, as the more answer texts the better, it seemed appropriate to constrain the grading to texts for which we had all relevant data (e.g. final grades). This was only an issue for one exam (i.e., Property, Instructor A) and there only for two students. After this fix was made, all results were rerun. I also edited the p value in text descriptions to reflect that shown in this running of the notebook. Measured to 1 significant figure the value didn't change, but to avoid confusion, I opted to have all references to the value match. Also, the code now references the `_redacted` versions of the data.